+++ app/types/response.go
@@ -4,8 +4,8 @@ package types
 type ProduceMessageResponseEntity struct {
 	Offset      int64       `json:"offset"`
 	Value       interface{} `json:"value"`
-	TopicId     uint64      `json:"topic_id"`
-	PartitionId uint64      `json:"partition_id"`
+	TopicId     uint        `json:"topic_id"`
+	PartitionId uint        `json:"partition_id"`
 }
 
 // Consumer Message Response Entity
